Removed namespace versioning + small refactor of settings + updated the unit tests
