+++ api/app/signals/__init__.py
@@ -31,7 +31,6 @@ datastructures.EmptyResultSet = EmptyResultSet
 VERSION = (2, 2, 0)
 
 API_VERSIONS = {
-    'v0': VERSION,
     'v1': VERSION,
 }
