+++ api/app/signals/apps/api/middleware.py
@@ -1,6 +1,6 @@
 # SPDX-License-Identifier: MPL-2.0
 # Copyright (C) 2018 - 2021 Gemeente Amsterdam
-from signals import API_VERSIONS
+from signals import VERSION
 from signals.utils.version import get_version
 
 
@@ -10,13 +10,15 @@ class APIVersionHeaderMiddleware:
         self.get_response = get_response
 
     def __call__(self, request):
+        # Versioning
+        # ==========
+        # SIA / Signalen follows the semantic versioning standard. Previously we had
+        # separate version numbers for the V0 (now defunct) and V1 versions of the API.
+        # We now no longer separately version these, as their releases were always
+        # tied to the backend. For backwards compatibility, and to not break external
+        # systems that rely on SIA / Signalen we still expose all the separate version
+        # numbers, but they are now all the same.
+
         response = self.get_response(request)
-
-        # Checking if we're coming from a DRF view, which contains API versioning information.
-        # If so, add a HTTP header with API version number.
-        renderer_context = getattr(response, 'renderer_context', {})
-        drf_request = renderer_context.get('request', None)
-        if drf_request and getattr(drf_request, 'version', None):
-            response['X-API-Version'] = get_version(API_VERSIONS[drf_request.version])
-
+        response['X-API-Version'] = get_version(VERSION)
         return response
