--- api/app/signals/apps/api/routers.py
@@ -1,68 +1,0 @@
-# SPDX-License-Identifier: MPL-2.0
-# Copyright (C) 2018 - 2021 Gemeente Amsterdam
-from rest_framework import routers
-from rest_framework_extensions.routers import ExtendedDefaultRouter
-
-
-class SignalsAPIRootViewVersion1(routers.APIRootView):
-    """Signalen API versie 1 (in development)."""
-
-    def get_view_name(self):
-        return 'Signals API Version 1'
-
-
-class SignalsRouterVersion1(ExtendedDefaultRouter):
-    APIRootView = SignalsAPIRootViewVersion1
-
-    # Overriding the `routes` attribute from the default DRF `routes`. We do this to control the
-    # usage of trailing slashes for different routes. In DRF you can only add or remove the
-    # trailing slash for all routes...
-    routes = [
-        # List route.
-        routers.Route(
-            url=r'^{prefix}{trailing_slash}$',
-            mapping={
-                'get': 'list',
-                'post': 'create',
-                'put': 'update',
-                'patch': 'partial_update',
-                'delete': 'destroy',
-            },
-            name='{basename}-list',
-            detail=False,
-            initkwargs={'suffix': 'List'}
-        ),
-        # Dynamically generated list routes. Generated using
-        # @action(detail=False) decorator on methods of the viewset.
-        routers.DynamicRoute(
-            url=r'^{prefix}/{url_path}{trailing_slash}$',
-            name='{basename}-{url_name}',
-            detail=False,
-            initkwargs={}
-        ),
-        # Detail route.
-        routers.Route(
-            url=r'^{prefix}/{lookup}{trailing_slash}$',
-            mapping={
-                'get': 'retrieve',
-                'put': 'update',
-                'patch': 'partial_update',
-                'delete': 'destroy'
-            },
-            name='{basename}-detail',
-            detail=True,
-            initkwargs={'suffix': 'Instance'}
-        ),
-        # Dynamically generated detail routes. Generated using
-        # @action(detail=True) decorator on methods of the viewset.
-        routers.DynamicRoute(
-            url=r'^{prefix}/{lookup}/{url_path}{trailing_slash}$',
-            name='{basename}-{url_name}',
-            detail=True,
-            initkwargs={}
-        ),
-    ]
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.trailing_slash = '/?'
