+++ api/app/signals/apps/api/templates/rest_framework/api.html
@@ -5,7 +5,7 @@
 {% block title %}Signals API{% endblock %}
 
 {% block branding %}
-    <a class='navbar-brand' rel="nofollow" href='{% url "v1:api-root" %}'>Signals API</a>
+    <a class='navbar-brand' rel="nofollow" href='{% url "api-root" %}'>Signals API</a>
 {% endblock %}
 
 {% block bootstrap_theme %}
