+++ api/app/signals/apps/questionnaires/templates/admin/questionnaires/questionnaire/change_form.html
@@ -2,7 +2,7 @@
 {% load i18n admin_urls %}
 {% block object-tools-items %}
     <li>
-        <a href="{% url 'v1:questionnaires:public-questionnaire-detail' uuid=original.uuid %}" class="historylink" target="_blank">Public API</a>
+        <a href="{% url 'questionnaires:public-questionnaire-detail' uuid=original.uuid %}" class="historylink" target="_blank">Public API</a>
     </li>
     {{block.super}}
 {% endblock %}
