+++ api/app/signals/apps/questionnaires/tests/rest_framework/views/test_private_questions_endpoint.py
@@ -14,7 +14,7 @@ THIS_DIR = os.path.dirname(__file__)
 
 
 urlpatterns = [
-    path('', include(('signals.apps.questionnaires.urls', 'signals.apps.questionnaires'), namespace='v1')),
+    path('', include('signals.apps.questionnaires.urls')),
 ]
