+++ api/app/signals/apps/questionnaires/tests/rest_framework/views/test_public_questions_endpoint.py
@@ -21,7 +21,7 @@ THIS_DIR = os.path.dirname(__file__)
 
 
 urlpatterns = [
-    path('', include(('signals.apps.questionnaires.urls', 'signals.apps.questionnaires'), namespace='v1')),
+    path('', include('signals.apps.questionnaires.urls')),
 ]
